DEV-101 git repository bootstrap with full application.
This is a backend service built with Spring Boot and Java 17 that fetches random facts from the Useless Facts API (https://uselessfacts.jsph.pl/random.json?language=en), provides a shortened URL for each fact, caches them, and offers a private area to consult access statistics of the shortened url. The service is designed with a focus on clean and maintainable architecture.